+++ lib/fast_mcp.rb
@@ -35,7 +35,7 @@ module FastMcp
   # @param options [Hash] Options for the middleware
   # @option options [String] :name The name of the server
   # @option options [String] :version The version of the server
-  # @option options [String] :path_prefix The path prefix for the MCP endpoints
+  # @option options [String] :path The path prefix for the MCP endpoints
   # @option options [Logger] :logger The logger to use
   # @yield [server] A block to configure the server
   # @yieldparam server [FastMcp::Server] The server to configure
@@ -117,7 +117,7 @@ module FastMcp
   # @param options [Hash] Options for the middleware
   # @option options [String] :name The name of the server
   # @option options [String] :version The version of the server
-  # @option options [String] :path_prefix The path prefix for the MCP endpoints
+  # @option options [String] :path The path prefix for the MCP endpoints
   # @option options [Logger] :logger The logger to use
   # @option options [Boolean] :authenticate Whether to use authentication
   # @option options [String] :auth_token The authentication token
@@ -129,7 +129,7 @@ module FastMcp
     name = options.delete(:name) || app.class.module_parent_name.underscore.dasherize
     version = options.delete(:version) || '1.0.0'
     logger = options[:logger] || Rails.logger
-    path_prefix = options.delete(:path_prefix) || '/mcp'
+    path = options.delete(:path) || '/mcp'
     authenticate = options.delete(:authenticate) || false
 
     options[:logger] = logger
@@ -145,7 +145,7 @@ module FastMcp
                                   end
 
     # Insert the middleware in the Rails middleware stack
-    app.middleware.use self.server.transport_klass, self.server, options.merge(path_prefix: path_prefix)
+    app.middleware.use self.server.transport_klass, self.server, options.merge(path: path)
   end
 
   # Notify the server that a resource has been updated